backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
torch==2.6.0
torchvision==0.21.0
numpy==2.2.4
matplotlib==3.10.1
pillow==11.1.0
tqdm
scikit-learn==1.6.1
adversarial-robustness-toolbox
requests==2.32.3